+++ packages/browser/src/client/utils.ts
@@ -1,0 +1,4 @@
+export const importId = (id: string) => {
+  const name = `/@id/${id}`
+  return import(name)
+}
